+++ src/main/java/org/infernalstudios/shieldexp/events/FovEvent.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright 2022 Infernal Studios
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.infernalstudios.shieldexp.events;
+
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.FOVUpdateEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import org.infernalstudios.shieldexp.ShieldExpansion;
+import org.infernalstudios.shieldexp.access.LivingEntityAccess;
+
+@Mod.EventBusSubscriber(modid = ShieldExpansion.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+public class FovEvent {
+    @SubscribeEvent
+    public void onFovModify(FOVUpdateEvent event) {
+        if (LivingEntityAccess.get(event.getEntity()).getBlocking()) event.setNewfov(1.0F);
+    }
+}
